100daysofcode - day 37 - learning grpc name-resolving
- Moved resolver into own package
- revisited the Builder and Resolver, and the UpdateState
- added comments
